+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - In `filterFasta`, when filter with expression table, changed to filter out peptides smaller than, instead of smaller or equal to, the value of `--quant-cutoff`.
 
+- Fixed the issue that in `splitFasta`, variant sources are not grouped as they are specified by `--group-source` #439
+
 ---
 
 ## [0.4.0] - 2022-03-17